--- a/requirements.txt
+++ b/requirements.txt
@@ -1,6 +1,7 @@
 langgraph>=0.2.0
 langchain-google-genai>=1.0.0
 google-cloud-bigquery>=3.13.0
+google-cloud-bigquery-storage>=2.0.0
 pandas>=2.0.0
 python-dotenv>=1.0.0 
 langchain_core>=0.3.0